Add ALU connections to CHERI ALU, general cleanup
This adds ALU connections to facilitate using the original ibex ALU
for adds in CHERI instructions in the future. Currently, this does not
work because we need the ALU to output a carry bit when adding.

This commit also cleans up some comments and sets some signals at the
start of the combinational ALU block.