Implement proper PCC checks, add MCCSR, MCAUSE, MEPCC
This commit implements PCC exception checks properly.
We add a new module to the toplevel in order to allow checking both
halfwords of the instruction properly - this is important because we
want to be able to execute the last instruction if it's compressed, even
if the second halfword that is read in is out of bounds.

We also pass the exceptions all the way through the fetch FIFO, because
we need to only check those exceptions when we're actually going to
execute the instruction - if we check them earlier then we might throw
an exception on an instruction that never gets executed.

The commit also implements the MCCSR CSR and MCAUSE & MEPCC SCRs