Add new sim file & update current simulation
This commit adds a new simulation file that reads in a .bin as the
initial memory for the simulation. It also has some miscellaneous code
for inputting commands into the core.
The commit also increases the size of the tag memory in order to be able
to run this new simulation file.
The old simulation file has been cleaned up, and a print statement has
been removed from socket_packet_utils.c