[dv/formal] DivInstrNotMult in separate step
DivInstrNotMult proves in Hp2 after MultEndState is proven, so I've
introduced an extra step for DivInstrNotMult and updated the verify TCL
script accordingly.

The proof for MultendState can be improved further by doing a graph
induction over the the multiplier and divide state machines. Essentially
when a multiply occurs, the divide must be idle and when a division is
happening the multiplier should be idle.